requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
redis==5.0.1
pyyaml==6.0.1
orjson==3.9.15
python-dotenv==1.0.1
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0